extract class MusicControl from GraphicalContext
